Moved OMDB properties into the properties package.
